Rename console output module
